requirements.txt
pandas>=1.5.0
google-generativeai>=0.3.0
tqdm>=4.65.0
openpyxl>=3.1.0
python-dotenv>=1.0.0 